chg: usr: Refactoring, moving
