+++ VERSION.txt
@@ -1,1 +1,1 @@
-v0.1.2
+v0.1.3
