+++ tests/unit/test_rule_utils.py
@@ -1,8 +1,8 @@
 from sklearn.tree import DecisionTreeRegressor
 from sklearn import datasets
-from rule.rule_utils import extract_rules_from_tree
-from condition.hyperrectanglecondition import HyperrectangleCondition
-from rule.rule import Rule
+from ruleskit.utils.rule_utils import extract_rules_from_tree
+from ruleskit.condition import HyperrectangleCondition
+from ruleskit.rule import Rule
 import numpy as np
 import pytest
